requirements.txt
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
cssselect==1.3.0
defusedxml==0.7.1
docx2pdf==0.1.8
filelock==3.20.0
Flask==3.1.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
parsel==1.10.0
pillow==11.3.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
python-docx==1.2.0
pywin32==311
queuelib==1.8.0
reportlab==4.4.4
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
tldextract==5.3.0
tqdm==4.67.1
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
Werkzeug==3.1.3
zope.interface==8.0.1
